settlement
updated

+++ api/settlement_files/delete_doc_info.php
@@ -3,16 +3,26 @@ require '../../ajaxconfig.php';
 
 $id = $_POST['id'];
 $result = 0;
-$qry = $pdo->query("SELECT upload FROM `document_info` WHERE id='$id'");
-if($qry->rowCount()>0){
-    $row = $qry->fetch();
-    unlink("../../uploads/doc_info/" . $row['upload']);
-    
-    $qry = $pdo->query("DELETE FROM `document_info` WHERE id='$id'");
-    if ($qry) {
-        $result = 1;
-    }else{
-        $result = 2;
+$cnt = '0';
+$qry = $pdo->query("SELECT * FROM `noc` WHERE id='$id");
+if ($qry->rowCount() > 0) {
+    $cnt = '1';
+}
+if ($cnt == '1') {
+    $result = '2'; // Used in User Creation.
+
+} else {
+    $qry = $pdo->query("SELECT upload FROM `document_info` WHERE id='$id'");
+    if ($qry->rowCount() > 0) {
+        $row = $qry->fetch();
+        unlink("../../uploads/doc_info/" . $row['upload']);
+
+        $qry = $pdo->query("DELETE FROM `document_info` WHERE id='$id'");
+        if ($qry) {
+            $result = 1;
+        } else {
+            $result = 2;
+        }
     }
 }
